+++ R/varpro_3_para.R
@@ -165,14 +165,9 @@ varpro_3_para_obj <- function(par, y, basis, t, nstart, sc_res = FALSE) {
   res
 }
 
-# following doc example gives an error:
-# @examples
-# varpro_3_para_opts(nstart = 10)
-
 #' Return a list of options for VARPRO based fitting with 3 free parameters:
-#' * zero'th order phase correction
-#' * global damping
-#' * global frequency shift.
+#' 1. zeroth order phase correction, 2. global damping, 3 global frequency
+#' shift.
 #' @param nstart position in the time-domain to start fitting, units of data
 #' points.
 #' @param init_damping starting value for the global Gaussian line-broadening